% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{AddNoise}
\alias{AddNoise}
\title{Add noise to response value}
\usage{
AddNoise(response.mat)
}
\arguments{
\item{response.mat}{A matrix. It contains the response data for one drug
combination.}
}
\value{
A matrix. It contains the response value added with noises.
}
\description{
Function \code{AddNoise} calculates and add a noise to values in response
matrix. The noises obey normal distribution ~N(0, 0.001) wich are generated
by fucntion \code{rnorm}.
}
\details{
\strong{Note}: If the analysis requires for reproductiblity, plesase set the
random seed before calling this function.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
response.mat <- data$dose.response.mats[[1]]
set.seed(1)
adjusted.mat <- AddNoise(response.mat)
}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
