% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape_data.R
\name{AdjustColumnName}
\alias{AdjustColumnName}
\title{Adjust column names of input data table}
\usage{
AdjustColumnName(data)
}
\arguments{
\item{data}{}
}
\value{
The data frame with the changed column names.
}
\description{
This function changes the column names in other format into the snake style.
i.e. block_id, drug_row, drug_col, .etc.
}
