% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct_plate.R
\name{BaselineCorrectionSD}
\alias{BaselineCorrectionSD}
\title{Baseline correction for the dose-response matrix of drug combinations}
\usage{
BaselineCorrectionSD(response.mat, fixed = c(NA, NA, NA, NA),
  nan.handle = c("LL4", "L4"))
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names.}

\item{fixed}{a parameter to specify which parameters are fixed and at what value they are fixed. NAs for parameter
that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
A list of the original dose-response matrix without correction and the corrected dose-response matrix.
}
\description{
A function to do baseline correction on the dose-response matrix for drug combinations with a weighted correction fator
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
data <- BaselineCorrectionSD(data$dose.response.mats[[1]])
}
\author{
Liye He \email{liye.he@helsinki.fi}, Jing Tang \email{jing.tang@helsinki.fi}
}

