% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CalculateBliss}
\alias{CalculateBliss}
\title{Calculate Bliss synergy score}
\usage{
CalculateBliss(response.mat)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}
}
\value{
A matrix for synergy score calculated via reference model introduced
by C. I. Bliss.
}
\description{
\code{CalculateBliss} calculates the synergy score matrix for a block of
drug combination by using a druginteraction reference model introduced by
C. I. Bliss in 1939.
}
\details{
This model is a reference model for evaluating the interaction between two
drugs. The basic assumption of this model is "The expected effect of two
drugs acting independently". More details about this model could be found in
original publication:
.
}
\references{
\itemize{
   \item{Yadav B, Wennerberg K, Aittokallio T, Tang J. (2015).
   \href{https://doi.org/10.1016/j.csbj.2015.09.001}{Searching for Drug
   Synergy in Complex Dose-Response Landscape Using an Interaction Potency
   Model.} Comput Struct Biotechnol J, 13:504– 513.}
   \item{Bliss, C. I. (1939).
   \href{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1744-7348.1939.tb06990.x}{The
   toxicity of poisons applied jointly.} Annals of Applied Biology,
   26(3):585–615.}
}
}
\author{
\itemize{
   \item{Liye He \email{liye.he@helsinki.fi}}
   \item{Shuyu Zheng \email{shuyu.zheng@helsinki.fi}}
}
}
