% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{CalculateDF}
\alias{CalculateDF}
\title{Calculate Drug Combination data in data frame format (template)}
\usage{
CalculateDF(template, debug = FALSE, method = "ZIP",
  data.type = "viability", ...)
}
\arguments{
\item{template}{a dataframe in the format as template. Columns "block_id",
"drug_row", "drug_col", "response", "conc_r", "conc_c", "conc_r_unit",
"conc_c_unit","cell_line_name", "drug_row", "drug_col" are reqired.}

\item{debug}{a logical value. If it is \code{TRUE}, block ID will be printed
to console. The default setting is \code{FALSE}}

\item{method}{a parameter to specify which models to use to calculate the 
synergy scores. Choices are "ZIP", "Bliss", "HSA" and "Loewe". Defaults to
"ZIP".}

\item{data.type}{a parameter to specify the response data type which can be 
either "viability" or "inhibition".}

\item{...}{Other arguments required by nested functions. Some important
arguments are:
 \itemize{
   \item \code{impute} and \code{noise} inherited from function
         \code{CalculateMat};
   \item \code{method} inherited from function \code{CorrectBaseLine};
   \item \code{Emin} and \code{Emax} inherited from function
         \code{FitDoseResponse}.
}}
}
\value{
a list. It contains 5 elements:
  \itemize{
    \item \strong{method} the method used to calculate the synergy scores.
    \item \strong{scores} It contains the modified response value and 4
      type of synergy scores of each drug dose response pair.
    \item \strong{dose.response.mats} The original input dose-response matrix.
    \item \strong{adjusted.response.mats} The dose response matrix adjusted 
      by functions: \code{\link{AddNoise}}, \code{\link{ImputeNear}}, and 
      \code{\link{CorrectBaseLine}}.
    \item \strong{drug.pairs} the same as the input data component.
  }
}
\description{
Calculate Drug Combination data in data frame format (template)
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
