% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{CalculateIC50}
\alias{CalculateIC50}
\title{Calculate Relative IC50 from Fitted Model}
\usage{
CalculateIC50(coef, type, max.conc)
}
\arguments{
\item{coef}{A numeric vector. It contains the fitted coefficients for
4-parameter log-logistic dose response model.}

\item{type}{A character value. It indicates the type of model was used for
fitting the dose-response curve. Available values are "L.4" and "LL.4".}

\item{max.conc}{A numeric value. It indicates the maximum concentration in
the dose-response data}
}
\value{
A numeric value. It is the relative IC50.
}
\description{
This function will calculate the relative IC50 from fitted 4-parameter 
log-logistic dose response model.
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
