% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pipeline.R
\name{CalculateMat}
\alias{CalculateMat}
\title{Main function for synergy score calculation from one dose-response matrix}
\usage{
CalculateMat(response.mat, noise = TRUE, method = "ZIP",
  data.type = "viability", correction = "non", ...)
}
\arguments{
\item{response.mat}{A matrix contain the drug combination reaponse value.
Its column names are doses of drug added along columns. Its row name are
doses of drug added along rows.}

\item{noise}{a logical value. It indicates whether or not adding noise to
to the "response" values in the matrix. Default is \code{TRUE}.}

\item{method}{a parameter to specify which models to use to calculate the 
synergy scores. Choices are "ZIP", "Bliss", "HSA" and "Loewe". Defaults to
"ZIP".}

\item{...}{Other argumants required by nested functions. Some important
arguments are:
\itemize{
   \item \code{method} inherited from function \code{\link{CorrectBaseLine}};
   \item \code{Emin} and \code{Emax} inherited from function
         \code{FitDoseResponse}.
}}
}
\value{
a list. It contains 4 elements:
  \itemize{
    \item \strong{method} the method used to calculate the synergy scores.
    \item \strong{scores} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{dose.response.mats} The original input dose-response matrix.
    \item \strong{adjusted.response.mats} The dose response matrix adjusted by
      functions: \code{\link{AddNoise}}, \code{\link{ImputeNear}}, and 
      \code{\link{CorrectBaseLine}}.
 }
}
\description{
\code{CalculateMat} is the main function for calculating synergy scores based
on model(ZIP, Bliss, Loewe, and HSA) fron one dose-response \strong{matrix}.
}
\details{
The steps for calculation:
\enumerate{
  \item Pre-process Matrix
    \enumerate{
      \item Impute for missing values (with the average of values from the 
          nearest four cells) in original matrix by using function 
          \code{\link{ImputeNear}}.
      \item Add noise(A small random number ranging from 0 to 0.001) to 
          original matrix by using function \code{line{AddNoise}}.
      )
      \item Correct baseline to 0, if \code{correction} is \code{TRUE}.
    }
  \item Calculate synergy scores using the functions:
    \itemize{
      \item \code{\link{CalculateZIP}}
      \item \code{\link{CalculateBliss}}
      \item \code{\link{CalculateHSA}}
      \item \code{\link{CalculateLoewe}}
    }
}
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
