% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{CalculateSensitivity}
\alias{CalculateSensitivity}
\title{Calculate the synergy scores for drug combinations}
\usage{
CalculateSensitivity(data, adjusted = TRUE, iteration = 100, seed = 123)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{adjusted}{a logical value. If it is \code{TRUE}, the
'adjusted.response.mats' will be used to calculate synergy scores. If it is
\code{FALSE}, the raw data ('dose.response.mats') will be used to calculate
synergy scores.}

\item{iteration}{a integer. It indicates the number of iterations for synergy
scores calculation on data with replicates.}

\item{seed}{a integer or NULL. It is used to set the random seed in synergy
scores calculation on data with replicates.}

\item{method}{a parameter to specify which models to use to calculate the
synergy scores. Choices are "ZIP", "Bliss", "HSA" and "Loewe". Defaults to
"ZIP".}
}
\value{
a list. It contains 4 elements:
  \itemize{
    \item \strong{dose.response.mats} The original input dose-response matrix
    \item \strong{adjusted.response.mats} The dose response matrix adjusted
      by functions: \code{\link{AddNoise}}, \code{\link{ImputeNA}}, and
      \code{\link{CorrectBaseLine}}.
    \item \strong{drug_pairs} a data frame contains the name of the row drug,
    the name of the column drug, concentration unit and block IDs.
    \item \strong{scores} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{method} the method used to calculate the synergy scores.
 }
}
\description{
\code{CalculateSynergy} is the main function for calculating synergy scores
based on model(ZIP, Bliss, Loewe, and HSA) fron one dose-response
\strong{matrix}.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
scores <- CalculateSensitivity(data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi}
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
 }

Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
