% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_score.R
\name{CalculateSynergy}
\alias{CalculateSynergy}
\title{Calculate the synergy scores for drug combinations}
\usage{
CalculateSynergy(data, method = "ZIP", correction = TRUE,
  correction.fixed = c(NA, NA, NA, NA), single.fixed = c(NA, NA, NA, NA),
  ZIP.Emax = NA, nan.handle = c("LL4", "L4"))
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}}

\item{method}{a parameter to specify which models to use to calculate the synergy scores. Choices are "ZIP",
"Bliss", "HSA" and "Loewe". Defaults to "ZIP".}

\item{correction}{a parameter to specify if baseline correction is used or not. Defaults to TRUE.}

\item{correction.fixed}{a parameter to specify which parameters of L.4 or LL.4 functions are fixed and at what value they are fixed for 
baseline correction. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{single.fixed}{a parameter to specify which parameters of L.4 or LL.4 functions are fixed and at what value they are fixed for 
single dose-response fitting. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}. It is only required by ZIP model.}

\item{ZIP.Emax}{the maximal inhibition of the drugs when applying ZIP model. By default, it is not fixed.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
a list of the following components:
\item{dose.response.mats}{ the same as the input data component.}
\item{drug.pairs}{the same as the input data component.}
\item{scores}{a list of matrixes of synergy scores.}
\item{method}{the method used to calculate the synergy scores.}
}
\description{
A function to calculate the synergy scores for drug combinations using different
models.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
scores <- CalculateSynergy(data)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}
\references{
Yadav B, Wennerberg K, Aittokallio T, Tang J. Searching for Drug Synergy in Complex Dose-Response Landscape Using an Interaction Potency Model.
Computational and Structural Biotechnology Journal 2015; 13: 504-513.
}

