% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CalculateZIP}
\alias{CalculateZIP}
\title{Calculate Delta synergy score based on ZIP model}
\usage{
CalculateZIP(response.mat, drug.row.model = NULL,
  drug.col.model = NULL, ...)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}

\item{drug.row.model}{(optional) a character. It indicates the model type
used for fitting dose-response curve for drug added to rows.}

\item{drug.col.model}{(optional) a character. It indicates the model used for
fitting dose-response curve for drug added to columns.}

\item{...}{Other arguments from nested functions.}
}
\value{
A matrix of \eqn{\Delta} score calculated via Zero Interaction
Potency (ZIP) method.
}
\description{
\code{CalculateZIP} calculates the \eqn{\Delta} score matrix for a block of
drug combination by using Zero Interaction Potency (ZIP) method.
}
\details{
zero interaction potency (ZIP) is a reference model for evaluating the
interaction between two drugs. It captures the drug interaction relationships
by comparing the change in the potency of the dose-response curves between
individual drugs and their combinations. More details about this model could
be found in original publication.
}
\references{
\itemize{
   \item{Yadav B, Wennerberg K, Aittokallio T, Tang J. (2015).
   \href{https://doi.org/10.1016/j.csbj.2015.09.001}{Searching for Drug
   Synergy in Complex Dose-Response Landscape Using an Interaction Potency
   Model.} Comput Struct Biotechnol J, 13:504– 513.}
}
}
\author{
\itemize{
   \item{Liye He \email{liye.he@helsinki.fi}}
   \item{Jing Tang \email{jing.tang@helsinki.fi}}
   \item{Shuyu Zheng \email{shuyu.zheng@helsinki.fi}}
}
}
