% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape_data.R
\name{.AdjustColumnName}
\alias{.AdjustColumnName}
\title{Adjust Column Names of Input Data Table}
\usage{
.AdjustColumnName(data)
}
\arguments{
\item{data}{A data frame. It is the input data for function 
\link{ReshapeData}}
}
\value{
The data frame with the changed column names.
}
\description{
This function changes the column names in other format into the style:
block_id, drug1, drug2, conc1, conc2, response, conc_unit1, conc_unit2.
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
