% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{.DimensionReduction}
\alias{.DimensionReduction}
\title{Dimension Reduction for Multi-drug Combination Visualization}
\usage{
.DimensionReduction(plot_table, drug_pair, distance_method)
}
\arguments{
\item{plot_table}{A data frame contains concentrations for all drugs, the
values for \code{plot_value}.}

\item{drug_pair}{A data frame contains the drug names and concentration
unites, whither the block is replicate or not.}

\item{distance_method}{The methods to calculate the distance between
different data points from the concentration of drugs. The distance matrix
is used for dimension reduction. This parameter is used to set the 
parameter \code{method} for \link[vegdist]{vegan}. The default values is
"euclidean".}
}
\value{

}
\description{
This function will take the multi-drug combination data, project the
concentrations of all the drugs into 2 dimension.
}
