% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_drugs.R
\name{.ExtendedScores}
\alias{.ExtendedScores}
\title{Make a Smooth Surface for Scores}
\usage{
.ExtendedScores(scores_mat, len)
}
\arguments{
\item{scores_mat}{a matrix contains scores which will be visualized}

\item{len}{length of the interval between plotted data points.}
}
\value{
a matrix which which contains interpolated points for input
  \code{scores_mat}.
}
\description{
Make a Smooth Surface for Scores
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
