% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{.Own_log}
\alias{.Own_log}
\title{CSS Facilitate Function - Log Calculation (nature based) LL.4 Model}
\usage{
.Own_log(b, c, x)
}
\arguments{
\item{b}{A numeric value. It is the fitted parameter b from 
\code{\link[drc]{L.4}} model.}

\item{c}{A numeric value. It is the fitted parameter c from 
\code{\link[drc]{L.4}} model.}

\item{x}{A numeric value. It is the relative IC50 for the curve. log10(e), 
where e is the fitted parameter e from \code{\link[drc]{L.4}} model.}
}
\value{
ln(1+10^(b*(c-x)))
}
\description{
#' This function calculates ln(1+10^(b*(c-x))) to be used in \code{.ScoreCurve}
function
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
