% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_drugs.R
\name{.Pt2mm}
\alias{.Pt2mm}
\title{Convert Font Size from pt to mm}
\usage{
.Pt2mm(x)
}
\arguments{
\item{x}{A numerical value. It is the font size in "pt" unite.}
}
\value{
A numerical value in "mm" unite
}
\description{
This function converts font sizes from "pt" unite to "mm" unite.
}
