% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_drugs.R
\name{.RoundValues}
\alias{.RoundValues}
\title{Round the Numbers for Plotting}
\usage{
.RoundValues(numbers)
}
\arguments{
\item{numbers}{A vector of numeric values. It contains the numbers need to
be rounded.}
}
\value{
A vector of rounded numbers.
}
\description{
This function will round the input numbers by 2 digits, if the absolute of 
number is larger than or equal to 1. It will take 2 significant digits, if
the absolute of number is smaller than 1.
}
