% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{.ScoreCurve_L4}
\alias{.ScoreCurve_L4}
\title{CSS Facilitate Function - .ScoreCurve for Curves Fitted by L.4 Model}
\usage{
.ScoreCurve_L4(b, c, d, e, c1, c2, t)
}
\arguments{
\item{b}{A numeric value, fitted parameter b from \code{\link[drc]{L.4}}
model.}

\item{c}{A numeric value, fitted parameter c from \code{\link[drc]{L.4}}
model.}

\item{d}{A numeric value, fitted parameter d from \code{\link[drc]{L.4}}
model.}

\item{e}{A numeric value, fitted parameter e from \code{\link[drc]{L.4}}
model.}

\item{c1}{A numeric value, log10(min conc) (this is the minimal nonzero
concentration).}

\item{c2}{A numeric value, log10(max conc) (this is the maximal
concentration).}

\item{t}{A numeric value, threshold (usually set to zero).}
}
\value{
A numeric value, RI or CSS scores.
}
\description{
This function is used to score sensitivities given either a single-agent or a
fixed conc (combination) columns. The function calculates the AUC of the
log10-scaled dose-response curve.
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
