% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{.SolveExpDoesLL4}
\alias{.SolveExpDoesLL4}
\title{Solve the Expected Dose of Drug to Achieve Given Effect (% inhibition)}
\usage{
.SolveExpDoesLL4(y, drug_par)
}
\arguments{
\item{y}{The expected effect (% inhibition) of the drug to cell line}

\item{drug_par}{The parameters for fitted dose-response model.}
}
\value{
A numeric value. It indicates the expected dose of drug.
}
\description{
This function will solve the fitted four-parameter logistic dose-response 
model and output the dose of drug at which it could achieve the % inhibition
to cell growth.
}
\author{
\itemize{
    \item Jing Tang \email{jing.tang@helsinki.fi}
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  }
}
