% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dose_response_model.R
\name{FindModelPar}
\alias{FindModelPar}
\title{Find the Fitted Parameters from 4-Parameter Log-Logistic Model}
\usage{
FindModelPar(model)
}
\arguments{
\item{model}{A object of class "drc".}
}
\value{
A numeric vector. It contains 4 fitted parameters.
}
\description{
Find the Fitted Parameters from 4-Parameter Log-Logistic Model
}
