% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{GenerateSurface}
\alias{GenerateSurface}
\title{3D Surface Plot for Nulti-drug Combination Dose-Response/Synergy Scores}
\usage{
GenerateSurface(
  dim_reduced_data,
  high_value_color,
  low_value_color,
  point_color,
  plot_title,
  legend_title,
  z_axis_title
)
}
\arguments{
\item{dim_reduced_data}{A list of data frame. It contains the dimension 
  reduced data for all the data points and other information for plotting. It
  is the output of \link{DimensionReduction}
(combination of concentrations). It is}

\item{high_value_color}{An R color value. It indicates the color for the
high values.}

\item{low_value_color}{An R color value. It indicates the color for low
values.}

\item{point_color}{An R color value. It indicates the color for data points.}

\item{plot_title}{A character value. It is the title for plot.}

\item{legend_title}{A character value. It is the title for legend.}

\item{z_axis_title}{A character value. It is the title for z-axis.}
}
\value{
A ggplot object.
}
\description{
This function will generate a surface plot for multi-drug combinations from
the output of \link{DimensionReduction}. It is an auxiliary function for 
\link{PlotMultiDrugSurface}
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
