% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{HighlightBarPlot}
\alias{HighlightBarPlot}
\title{Highlight Bars}
\usage{
HighlightBarPlot(selected_data)
}
\arguments{
\item{selected_data}{A data frame. It contain the information for the bars
to be highlighted.}
}
\value{
A ggplot object
}
\description{
It is an auxiliary function for \link{PlotMultiDrugBar}
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
