% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA.R
\name{HSA}
\alias{HSA}
\title{HSA synergy score based on highest single agent (HSA) model}
\usage{
HSA(response.mat, correction = TRUE, correction.fixed = c(NA, NA, NA, NA),
  nan.handle = c("LL4", "L4"))
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names}

\item{correction}{a parameter to specify if baseline correction is used or not. Defaults to TRUE.}

\item{correction.fixed}{a parameter to specify which parameters of L.4 or LL.4 functions are fixed and at what value they are fixed for 
baseline correction. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
A matrix of HSA synergy scores for all the dose pairs for a drug combination. For a
does pair with at least one zero concentration, 0 is used as the synergy score.
}
\description{
A function to calculate HSA synergy score based on HSA model
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
delta.score <- HSA(data$dose.response.mats[[1]])
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

