% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape_data.R
\name{ImputeNA}
\alias{ImputeNA}
\title{Impute missing value with nearest values}
\usage{
ImputeNA(response.mat)
}
\arguments{
\item{response.mat}{A matrix which has missing value.}

\item{times}{a integer which refers to the time of runing imputation. The
default value is \code{1}, but sometimes you might want to run imputation
more than once.}
}
\value{
A matrix which is same as input matrix except the NA cells are filled
with numbers.
}
\description{
Function \code{ImputeNA} do missing value imputation by assigning the
average of values in nearest 4 cells (top, bottom, left, right) to the NA
cell.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
response.mat <- data$dose.response.mats[[1]]
# introduce some NA values into matrix
response.mat[3:4, 3:5] <- NA
adjusted.mat <- ImputeNa(response.mat)
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
