% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_respose_matrix.R
\name{ImputeNear}
\alias{ImputeNear}
\title{Impute missing value with nearest values}
\usage{
ImputeNear(response.mat, times = 1)
}
\arguments{
\item{response.mat}{A matrix which has missing value.}

\item{times}{a integer which refers to the time of runing imputation. The
default value is \code{1}, but sometimes you might want to run imputation
more than once.}
}
\value{
A matrix which is same as input matrix except the NA cells are filled
with numbers.
}
\description{
Function \code{ImputeNear} do missing value inputation by assigning the
average of values in nearest 4 cells (top, bottom, left, right) to the NA
cell.
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
