% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging}
\alias{kriging}
\title{Kriging}
\usage{
kriging(
  data,
  data.coord,
  krig.coord,
  cov.mod = "whitmat",
  sill,
  range,
  smooth,
  smooth2 = NULL,
  grid = FALSE,
  only.weights = FALSE
)
}
\arguments{
\item{data}{A numeric vector or matrix. If data is a matrix then the simple
kriging predictor is given for each realisation, i.e., each row of data.}

\item{data.coord}{A numeric vector or matrix specifying the coordinates of
the observed data. If data.coord is a matrix, each row must corresponds to
one location.}

\item{krig.coord}{A numeric vector or matrix specifying the coordinates where
the kriging predictor has to be computed. If krig.coord is a matrix, each
row must correspond to one location.}

\item{cov.mod}{A character string specifying the covariance function family.
Must be one of "whitmat", "powexp", "cauchy", "bessel" or "caugen" for the
Whittle-Matern, the powered exponential, the Cauchy, the Bessel or the
generalized Cauchy covariance families.}

\item{sill, range, smooth, smooth2}{Numerics specifiying the sill, range, smooth
and, if any, the second smooth parameters of the covariance function.}

\item{grid}{logical. Does krig.coord specifies a grid?}

\item{only.weights}{Logical. Should only the kriging weights be computed? If
FALSE, the kriging predictor isn't computed.}
}
\value{
A list with components
\itemize{
  \item coord	The coordinates where the kriging predictor has been computed;
  \item krig.est	The kriging predictor estimates;
  \item grid	Does coord define a grid?;
  \item weights A matrix giving the kriging weights: each column corresponds
  to one prediction location.
}
}
\description{
This function interpolates a zero mean Gaussian random field using the simple
kriging predictor.
}
