% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCATS_10023_data.R
\name{NCATS_10023_data}
\alias{NCATS_10023_data}
\title{A high-throughput 3 drug combination screening data}
\format{
A data frame with the following columns: BlockId, Drug1, Drug2,
  Drug3, Conc1, Conc2, Conc3, Response..
}
\description{
A 3-drug combination screening data on Malaria. It is downloaded from 
[NCATS Matrix](https://matrix.ncats.nih.gov/) project 2321 "Malaria TACT",
assay 10023.
}
