% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCATS_screening_data.R
\name{NCATS_screening_data}
\alias{NCATS_screening_data}
\title{A high-throughput 3 drug combination screening data}
\format{
A data frame with the following columns: "block_id", "drug1",
  "drug2", "drug3", "conc1", "conc2", "conc3", "response", "conc_unit1",
  "conc_unit2", "conc_unit3"
}
\description{
A 3-drug combination screening data on Malaria. It is downloaded from 
[NCATS Matrix](https://matrix.ncats.nih.gov/) project 2321 "Malaria TACT".
It contains 2 blocks, one is synergistic and the other is antagonistic.
}
