% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity.R
\name{own_log}
\alias{own_log}
\title{CSS facilitate function - log calculation (nature based) LL.4 version}
\usage{
own_log(b, c, x)
}
\arguments{
\item{b}{fitted parameter b from \code{\link[drc]{L.4()}} model}

\item{c}{fitted parameter c from \code{\link[drc]{L.4()}} model}

\item{x}{relative IC50 for the curve. log10(e), where e is the
fitted parameter e from \code{\link[drc]{L.4()}} model}
}
\value{
ln(1+10^(b*(c-x)))
}
\description{
#' This function calculates ln(1+10^(b*(c-x))) to be used in \code{scoreCurve}
function
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
