% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity.R
\name{own_log2}
\alias{own_log2}
\title{CSS facilitate function - log (nature based) calculation L.4 version}
\usage{
own_log2(x)
}
\arguments{
\item{x}{relative IC50 for the curve. The fitted parameter e from
\code{\link[drc]{L.4()}} model}
}
\value{
ln(1+exp(x))
}
\description{
This function calculates ln(1+exp(x)) to be used in \link{scoreCurve.L4}
function
}
\author{
Jing Tang \email{jing.tang@helsinki.fi}
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
