% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_dose_response.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the drug combination dose-response data}
\usage{
PlotDoseResponse(data, save.file = FALSE, pair.index = NULL,
  single.fixed = c(NA, NA, NA, NA), type = c("both", "heatmap", "curve"),
  merge = FALSE, pdf.height = 8, pdf.width = 8, ...)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{save.file}{a parameter to specify if the visualization results are saved as pdf files in current working directory or not. If it is FALSE, the results are
returned as a list of the plots. It is FALSE by default.}

\item{pair.index}{a parameter to specify which drug combination if there are many drug combinations in the data.
By default, it is NULL so that the visualization of all the drug combinations in the data is returned.}

\item{single.fixed}{a parameter to specify which parameters are fixed and at what value they are fixed when applying 4 parameters
log-logistic function to fit single drug response. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{type}{a parameter to specify the types of the plot. There are three options: "curve", "heatmap" and "both". If it is "curve", only dose-response
curves for single drugs are plotted. If it is "heatmap", the dose-response matrix will be plotted as a heatmap. If it is "both", both plots are returned.}

\item{merge}{a parameter to specify if the two dose-response curves are merged into one plot or not. It only works when type is set to curve.}

\item{pdf.height}{a parameter to specify the height of the pdf file when save.file is TRUE. Defaults to 8.}

\item{pdf.width}{a parameter to specify the width of the pdf file when save.file is TRUE. Defaults to 8.}

\item{...}{further graphical parameters from \code{\link{plot}} for plotting the single drug dose-response curve. 
Use e.g., cex.lab to change the axis label size and cex.axis to change the tick size of axises.}
}
\value{
if save.file parameter is TRUE, pdf files are returned. Otherwise, the plots are only displayed.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponse(data, type = "both", save.file = TRUE)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

