% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dose_response_curve.R
\name{PlotDoseResponseCurve}
\alias{PlotDoseResponseCurve}
\title{Plot Dose Response Curve for Single Drug}
\usage{
PlotDoseResponseCurve(
  data,
  plot_block = 1,
  drug_index = 1,
  adjusted = TRUE,
  Emin = NA,
  Emax = NA,
  grid = expression(graphics::grid(col = "#DFDFDF", lty = 1)),
  point_color = "#C24B40",
  curve_color = "black",
  plot_setting = list(cex.lab = 1, cex.axis = 1, mgp = c(2, 0.5, 0), font.main = 2,
    font.lab = 3, cex.main = 14/12, bty = "l", lwd = 1.5)
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{plot_block}{A character/integer. It indicates the block ID for the
block to visualize.}

\item{drug_index}{A character/integer. It indicates the index of the drug to 
  plot. For example, 1 or "1" indicates to plot "drug1" in the input 
\code{data}.}

\item{adjusted}{A logical value. If it is \code{FALSE}, original response
value will be used to fit the curve. If it is \code{TRUE}, the response
values adjusted by (adding noise and/or imputation) will be plotted.}

\item{Emin}{A numeric or \code{NA}. the minimal effect of the drug used in
the 4-parameter log-logistic function to fit the dose-response curve. If
it is not NA, it is fixed the value assigned by the user. Default setting
is \code{NA}.}

\item{Emax}{A numeric or \code{NA}. the maximal effect of the drug used in
the 4-parameter log-logistic function to fit the dose-response curve. If
it is not NA, it is fixed the value assigned by the user. Default setting
is \code{NA}.}

\item{grid}{A expression for \link[graphics]{grid} function of \code{NULL}.
If it is \code{NULL}, no grids will be shown in the plot.}

\item{point_color}{An R color value. It indicates the color for points in
dose response curve plots.}

\item{curve_color}{An R color value. It indicates the color for curves in 
dose response curve plots.}

\item{plot_setting}{A list of graphical arguments. The arguments are passed 
to \link[graphics]{par} function to modify the appearance of plots.}
}
\value{
A plot object recorded by \link[grDevices]{recordPlot}.
}
\description{
This function will pot the dose response curve fitted by 4 parameters 
log-logistic curve.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponseCurve(data, grid = NULL)
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
