% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{PlotMultiDrugBar}
\alias{PlotMultiDrugBar}
\title{Bar Plot for Multi-drug Combination Dose-Response/Synergy Scores}
\usage{
PlotMultiDrugBar(
  data,
  plot_block = 1,
  plot_value = "response",
  sort_by = "conc1",
  highlight_row = NULL,
  bar_color = "grey50",
  highlight_color = "#A90217"
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{plot_block}{A character/integer. It indicates the block ID for the
block to visualize.}

\item{plot_value}{A vector of characters. It contains the name of one or more
metrics to be visualized. If the \code{data} is the direct output from 
\link{ReshapeData}, the values for this parameter are:
\itemze{
  \item \strong{response_origin} The original response value in input data.
  It might be % inhibition or % viability.
  \item \strong{response} The % inhibition after preprocess by function 
  \link{ReshapeData}
}
If the \code{data} is the output from \link{CalculateSynergy}, following
values are also available:
\itemze{
  \item \strong{ZIP_ref, Bliss_ref, HSA_ref, Loewe_ref} The reference
  additive effects predicted by ZIP, Bliss, HSA or Loewe model,
  respectively.
  \item \strong{ZIP_synergy, Bliss_synergy, HSA_synergy, Loewe_synergy}
  The synergy score calculated by ZIP, Bliss, HSA or Loewe model,
  respectively.
  \item \strong{ZIP_fit} The response fitted by ZIP model.
}}

\item{sort_by}{A character. It indicates by which metric the bars (data 
points) will be sorted. It could be one of the available values for
\code{plot_value} or one of the concentration columns (e.g. "cocn1",
"conc2", ...)}

\item{highlight_row}{A vector of numeric values with the length same as the
number of drugs in selected block. It contains the concentrations  for
"drug1", "drug2", ... The data point selected by these concentrations will
be highlighted in the plot.}

\item{bar_color}{An R color value. It indicates the color of bars.}

\item{highlight_color}{An R color value. It indicates the color of
highlighted bar.}
}
\value{
A ggplot object.
}
\description{
This function will generate a group of bar plots for one drug combination
block. Each panel (columns) visualize the concentrations for all the drugs 
and metrics specified by \data{plot_values}. Each row represents a data point
in the combination data. The data point specified by \code{highlight_row}
will be highlighted in different color.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotMultiDrugBar(data, highlight_row = c(0, 0))
}
