% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitive_synergy.R
\name{PlotSensitiveSynergy}
\alias{PlotSensitiveSynergy}
\title{Plot Sensitive-Synergy Plot for All the Combinations in the Input Data}
\usage{
PlotSensitiveSynergy(
  data,
  plot_synergy,
  point_color = "#2D72AD",
  point_label_color = "#2D72AD"
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{plot_synergy}{A character value. It indicates the synergy score for
visualization. The available values are: "ZIP", "HSA", Bliss", "Leowe".}

\item{point_color}{An R color value. It indicates the color for the points.}

\item{point_label_color}{An R color value. It indicates the color for the
label of data points.}
}
\value{
A ggplot object.
}
\description{
This function will generate a scatter plot for all the combinations in the
input data. The x-axis is the Combination Sensitive score (CSS).
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
data <- CalculateSynergy(data, method = c("ZIP"))
data <- CalculateSensitivity(data)
PlotSensitiveSynergy(data, plot_synergy = "ZIP")
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
