% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{ReshapeData}
\alias{ReshapeData}
\title{Pre-process the response data for further calculation and plot}
\usage{
ReshapeData(
  data,
  impute = TRUE,
  impute_method = NULL,
  noise = TRUE,
  seed = NULL,
  data_type = "viability"
)
}
\arguments{
\item{data}{drug combination response data in a data frame format}

\item{impute}{a logical value. If it is \code{TRUE}, the \code{NA} values
will be imputed by \code{\link[mice]{mice}}. Default is \code{TRUE}.}

\item{impute_method}{a single string. It sets the \code{method} parameter
in function \code{\link[mice]{mice}} to specify the imputation method. 
Please check the documentation of \code{\link[mice]{mice}} to find the
available methods.}

\item{noise}{a logical value. It indicates whether or not adding noise to
to the "response" values in the matrix. Default is \code{TRUE}.}

\item{seed}{a single value, interpreted as an integer, or NULL. It is the
random seed for calculating the noise. Default setting is \code{NULL}}

\item{data_type}{a parameter to specify the response data type which can be
either "viability" or "inhibition".}

\item{correction}{a character. This argument is extended from the argument
\code{method} of \code{\link{CorrectBaseLine}} function. There are three
available valuse: \code{non}, \code{part}, \code{all}.
The default setting is \code{non}.}
}
\value{
a list of the following components:
  \itemize{
    \item \strong{drug_pairs} A data frame contains the name of all the
    tested drugs, concentration unit, block IDs and a logical column 
    "replicate" to indicate whether there are replicates in the corresponding
    block.
    \item \strong{response} A data frame contains the columns: "concX" 
    concentrations for drugs from input data; "response_origin" response
    values from input data; "response" % inhibition value for downstream
    analysis.
  \item \strong{response_statistics} A data frame. It will be output if
    there is block have replicated response values. It contains the
    block ID, the concentrations for all the tested drugs, and statistics for
    % inhibition values across replicates (including mean, standard
    deviation, standard error of mean and 95% confidence interval).
  }
}
\description{
A function to transform the response data from data frame format to
dose-response matrices. Several processes could be chose to add noise, impute
missing values or correct base line to the dose-response matrix.
}
\details{
The input data must contain the following columns:
(block_id/BlockId/PairIndex), (drug_row/DrugRow/Drug1),
(drug_col/DrugCol/Drug2), (response/Response), (conc_r/ConcRow/Conc1),
(conc_c/ConcCol/Conc2), and (ConcUnit/conc_r_unit, conc_c_unit/ConcUnit1,
ConcUnit2, ConcUnit3)
}
\examples{
data("mathews_screening_data")
# set a random number seed for generating the noises
set.seed(1)
data <- ReshapeData(mathews_screening_data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi}
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
 }
}
