% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity.R
\name{RIConfidenceInterval}
\alias{RIConfidenceInterval}
\title{Calculate Confidence Interval of RI}
\usage{
RIConfidenceInterval(pred, iter = 100)
}
\arguments{
\item{pred}{A data frame. It must contain:
  \itemize{
    \item \strong{dose} The concentration of drugs used for model fitting
    and prediction.
    \item \strong{prediction} The predicted response value at certain
    doses.
    \item \strong{sd} The standard deviation of predicted response value.
  }
It can be generated by calling \code{\link{CalculateSens}(df, pred = TRUE)}.}

\item{iter}{A numeric value. It indicates the number of iterations for
simulation.}
}
\value{
A named numeric vector. It contains simulated RI, two boundaries of
the RI's confidence interval
}
\description{
Function \code{RIConfidenceInterval} is used to calculate The confidence
interval of RI (Relitave Inhibition) score for cell sensitivity to certain
drug.
}
\details{
\code{RIConfidenceInterval} takes the prediction results from
\code{\link{CalculateSens}} to generate simmulated dose respons data. The
number of iteration is controlled by parameter \code{iter}. The RIs will
be computated for all the simulated data. Finally, the function will return:
Simulated RI (Median), and two boundaries of RI's confidence interval.
}
\examples{
df <- data.frame(dose = c(0, 0.1954, 0.7812, 3.125, 12.5, 50),
                 response = c(2.95, 3.76, 18.13, 28.69, 46.66, 58.82))

RI_with_pred <- CalculateSens(df, pred = TRUE)

}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
Jing Tang \email{jing.tang@helsinki.fi}
}
