% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{ZIP}
\alias{ZIP}
\title{Delta synergy score based on zero interaction potency (ZIP) model}
\usage{
ZIP(response.mat, correction = TRUE, correction.fixed = c(NA, NA, NA, NA),
  single.fixed = c(NA, NA, NA, NA), ZIP.Emax = NA, nan.handle = c("LL4",
  "L4"))
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names}

\item{correction}{a parameter to specify if the baseline correction is used or not. Defaults to TRUE.}

\item{correction.fixed}{a parameter to specify which parameters of L.4 or LL.4 functions are fixed and at what value they are fixed for 
baseline correction. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{single.fixed}{a parameter to specify which parameters of L.4 or LL.4 functions are fixed and at what value they are fixed for 
single dose-response fitting. NAs for parameter that are not fixed. More details in function \code{\link[drc]{drm}}.}

\item{ZIP.Emax}{the maximal inhibition of the drugs when applying ZIP model. By default, it is not fixed.}

\item{nan.handle}{a parameter to specify if L.4 function or LL.4 function is used when fitting with LL.4 produces
NaNs.}
}
\value{
A matrix of delta scores for all the dose pairs for a drug combination. For a does pair with at least one zero concentration, 0 is used as the synergy score.
}
\description{
A function to calculate delta synergy score based on zero interaction potency (ZIP) model
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
delta.score <- ZIP(data$dose.response.mats[[1]])
}
\author{
Liye He \email{liye.he@helsinki.fi}, Jing Tang \email{jing.tang@helsinki.fi}
}
\references{
Yadav B, Wennerberg K, Aittokallio T, Tang J. Searching for Drug Synergy in Complex Dose-Response Landscape Using an Interaction Potency Model.
Computational and Structural Biotechnology Journal 2015; 13: 504-513.
}

