% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{AdjustColumnName}
\alias{AdjustColumnName}
\title{Adjust column names of input data table}
\usage{
AdjustColumnName(data)
}
\arguments{
\item{data}{}
}
\value{
The data frame with the changed column names.
}
\description{
This function changes the column names in other format into the style:
block_id, drug1, drug2, conc1, conc2, response, conc_unit1, conc_unit2.
}
