% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct_plate.R
\name{BaselineCorrectionSD}
\alias{BaselineCorrectionSD}
\title{Baseline correction for the dose-response matrix of drug combinations}
\usage{
BaselineCorrectionSD(response.mat)
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names}
}
\value{
A list of the original dose-response matrix without correction and the corrected dose-response matrix
}
\description{
A function to do baseline correction on the dose-response matrix for drug combinations with a weighted correction fator
}
\author{
Liye He \email{liye.he@helsinki.fi}, Jing Tang \email{jing.tang@helsinki.fi}
}

