% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{CalculateCSS}
\alias{CalculateCSS}
\title{Calculate combination sensitivity score}
\usage{
CalculateCSS(response, ic50)
}
\arguments{
\item{response}{A data frame. It must contain the columns: "conc1", "conc2",
..., for the concentration of the combined drugs and "response" for the
observed %inhibition at certain combination.}

\item{ic50}{A list. It contains the relative IC50 for all the drugs in the
combination.}
}
\value{
A data frame. It contains the CSS for each drug will one of the other
  drugs is at its IC50 and summarized CSS for the whole block.
}
\description{
This function will calculate the Combination Sensitivity Score (CSS) for a
drug combination block.
}
