% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity_score.R
\name{CalculateSensitivity}
\alias{CalculateSensitivity}
\title{Calculate the Synergy Scores for Drug Combinations}
\usage{
CalculateSensitivity(data, adjusted = TRUE, iteration = 10, seed = 123)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{adjusted}{A logical value. If it is \code{TRUE}, the
'adjusted.response.mats' will be used to calculate synergy scores. If it is
\code{FALSE}, the raw data ('dose.response.mats') will be used to calculate
synergy scores.}

\item{iteration}{An integer. It indicates the number of iterations for
synergy scores calculation on data with replicates.}

\item{seed}{An integer or NULL. It is used to set the random seed in synergy
scores calculation on data with replicates.}
}
\value{
This function will add columns into \code{data$drug_pairs} table.
  The columns are:
  \itemize{
    \item \strong{ic50_1/2/...} Relative IC50 for drug 1, 2, ...
    \item \strong{ri_1/2/...} Relative Inhibition (RI) for drug 1, 2, ...
    \item \strong{css1_ic502/...} CSS score of drug 1 while fixing drug 2
      at its IC50.
    \item \strong{css} Over all CSS score for the whole block. It's the mean
      value of the CSS for all drug pairs in the combination. 
 }
 If there are replicates in the block, this function will add one table named
 as "sensitivity_scores_statistics" for the statistics of the values
 mentioned about into the input \code{data} list.
}
\description{
\code{CalculateSynergy} is the main function for calculating synergy scores
based on model(ZIP, Bliss, Loewe, and HSA) fron one dose-response
\strong{matrix}.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
data <- CalculateSensitivity(data)
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
