% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{CalculateSynergy}
\alias{CalculateSynergy}
\title{Calculate the synergy scores for drug combinations}
\usage{
CalculateSynergy(data, method = "ZIP", adjusted = TRUE,
  data.type = "viability")
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{method}{a parameter to specify which models to use to calculate the 
synergy scores. Choices are "ZIP", "Bliss", "HSA" and "Loewe". Defaults to
"ZIP".}

\item{adjusted}{a logical value. If it is \code{TRUE}, the 
'adjusted.response.mats' will be used to calculate synergy scores. If it is
\code{FALSE}, the raw data ('dose.response.mats') will be used to calculate
synergy scores.}

\item{data.type}{a parameter to specify the response data type which can be 
either "viability" or "inhibition".}
}
\value{
a list. It contains 4 elements:
  \itemize{
    \item \strong{dose.response.mats} The original input dose-response matrix.
    \item \strong{adjusted.response.mats} The dose response matrix adjusted by
      functions: \code{\link{AddNoise}}, \code{\link{ImputeNear}}, and 
      \code{\link{CorrectBaseLine}}.
    \item \strong{drug.pairs} a data frame contains the name of the row drug,
    the name of the column drug, concentration unit and block IDs.
    \item \strong{scores} It contains the modified response value and 4
    type of synergy scores of each drug dose response pair.
    \item \strong{method} the method used to calculate the synergy scores.
 }
}
\description{
\code{CalculateSynergy} is the main function for calculating synergy scores based
on model(ZIP, Bliss, Loewe, and HSA) fron one dose-response \strong{matrix}.
}
\details{
The steps for calculation:
\enumerate{
  \item Pre-process Matrix
    \enumerate{
      \item Impute for missing values (with the average of values from the 
          nearest four cells) in original matrix by using function 
          \code{\link{ImputeNear}}.
      \item Add noise(A small random number ranging from 0 to 0.001) to 
          original matrix by using function \code{line{AddNoise}}.
      )
      \item Correct baseline to 0, if \code{correction} is \code{TRUE}.
    }
  \item Calculate synergy scores using the functions:
    \itemize{
      \item \code{\link{ZIP}}
      \item \code{\link{Bliss}}
      \item \code{\link{HSA}}
      \item \code{\link{Loewe}}
    }
}
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
scores <- CalculateSynergy(data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi}
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
 }
}
