% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_synergy.R
\name{.ExtendedScores}
\alias{.ExtendedScores}
\title{Make a smooth surface for scores}
\usage{
.ExtendedScores(scores.mat, len)
}
\arguments{
\item{scores.mat}{a matrix contains scores which will be visualized}

\item{len}{length of the interval between plotted data points.}
}
\value{
a matrix which
}
\description{
Make a smooth surface for scores
}
