% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{ExtractSingleDrug}
\alias{ExtractSingleDrug}
\title{Extract single drug response from matrix}
\usage{
ExtractSingleDrug(response.df)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column
and row, respectively. The values in matrix indicate the inhibition rate to
cell growth.}
}
\value{
A list contains several data frames each of which contains two columns:
  \itemize{
    \item \strong{dose} The concertration of drug.
    \item \strong{response} The cell's response (inhibation rate) to
      corresponding drug concertration.
}
}
\description{
\code{ExtractSingleDrug} extracts the dose-response values of single drug
 from a drug combination dose-response matrix.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
response.mat <- data$dose.response.mats[[1]]
drug.row <- ExtractSingleDrug(response.df)
}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
