% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_respose_matrix.R
\name{ExtractSingleDrug}
\alias{ExtractSingleDrug}
\title{Extract single drug response from matrix}
\usage{
ExtractSingleDrug(response.mat, dim = "row")
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column and
row, respectively. The values in matrix indicate the inhibition rate to cell
growth.}

\item{dim}{A character. It should be either "col" or "row" to indicate which
drug's dose-response value will be extracted.}
}
\value{
A data frame. It contains two variables:
\itemize{
\item \strong{dose} The concertration of drug.
\item \strong{response} The cell's response (inhibation rate) to
corresponding drug concertration.
}
}
\description{
\code{ExtractSingleDrug} extracts the dose-response values of single drug (
drug added in column or row) from a drug combination dose-response matrix.
}
\author{
Shuyu Zheng{shuyu.zheng@helsinki.fi}
}
