% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{GenerateSurface}
\alias{GenerateSurface}
\title{3D Surface Plot for Nulti-drug Combination Dose-Response/Synergy Scores}
\usage{
GenerateSurface(
  dim_reduced_data,
  high_value_color,
  low_value_color,
  show_data_points = TRUE,
  point_color,
  plot_title,
  plot_subtitle,
  legend_title,
  z_axis_title,
  text_size_scale = 1,
  camera_width = NULL,
  camera_height = NULL,
  camera_scale = 1
)
}
\arguments{
\item{dim_reduced_data}{A list of data frame. It contains the dimension 
  reduced data for all the data points and other information for plotting. It
  is the output of \link{DimensionReduction}
(combination of concentrations). It is}

\item{high_value_color}{An R color value. It indicates the color for the
high values.}

\item{low_value_color}{An R color value. It indicates the color for low
values.}

\item{show_data_points}{A logical value. If it is \code{TRUE}, the raw data
points will be shown on the plot. If it is \code{FALSE}, no points will be
plotted.}

\item{point_color}{An R color value. It indicates the color for data points.}

\item{plot_title}{A character value. It is the title for plot.}

\item{plot_subtitle}{A character value. It is the subtitle for plot.}

\item{legend_title}{A character value. It is the title for legend.}

\item{z_axis_title}{A character value. It is the title for z-axis.}

\item{text_size_scale}{A numeric value. It is used to control the size
of text in the plot. All the text size will multiply by this scale factor.}

\item{camera_width}{A numeric value or NULL. It indicates the output figure's
width on pixel.}

\item{camera_height}{A numeric value or NULL. It indicates the output
figure's height on pixel.}

\item{camera_scale}{A numeric value. The output plot will multiply 
title/legend/axis/canvas sizes by this factor.}
}
\value{
A ggplot object.
}
\description{
This function will generate a surface plot for multi-drug combinations from
the output of \link{DimensionReduction}. It is an auxiliary function for 
\link{PlotMultiDrugSurface}
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
