% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{HighlightBarPlot}
\alias{HighlightBarPlot}
\title{Highlight Bars}
\usage{
HighlightBarPlot(selected_data, text_size = 10)
}
\arguments{
\item{selected_data}{A data frame. It contain the information for the bars
to be highlighted.}

\item{text_size}{A numeric value. It indicates the label text size in "pt"
for the highlighted row.}
}
\value{
A ggplot object
}
\description{
It is an auxiliary function for \link{PlotMultiDrugBar}
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
