% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA.R
\name{HSA}
\alias{HSA}
\title{HSA synergy score based on highest single agent (HSA) model}
\usage{
HSA(response.mat, correction = TRUE)
}
\arguments{
\item{response.mat}{a dose-response matrix with concentrations as row names and column names}

\item{correction}{a parameter to specify if baseline correction is used or not. Defaults to TRUE.}
}
\value{
A matrix of HSA synergy scores for all the dose pairs for a drug combination. For a
does pair with at least one zero concentration, 0 is used as the synergy score.
}
\description{
A function to calculate HSA synergy score based on HSA model
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

