% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape_data.R
\name{ImputeNA}
\alias{ImputeNA}
\title{Impute missing value with nearest values}
\usage{
ImputeNA(response.mat)
}
\arguments{
\item{response.mat}{A matrix which has missing value.}
}
\value{
A matrix which is same as input matrix except the missing values are
imputed.
}
\description{
Function \code{ImputeNA} does missing value imputation by assigning the
average of values in nearest 4 cells (top, bottom, left, right) to the NA
cell. This pocess will be done repeadly until there is no missing values in
the matrix.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
response.mat <- data$dose.response.mats[[1]]
# introduce some NA values into matrix
response.mat[3:4, 3:5] <- NA
adjusted.mat <- ImputeNA(response.mat)
}
\author{
Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
}
