% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_synergy_score.R
\name{Loewe}
\alias{Loewe}
\title{Calculate Loewe synergy score}
\usage{
Loewe(response.mat, quiet = TRUE, drug.col.model = NULL,
  drug.row.model = NULL)
}
\arguments{
\item{response.mat}{A drug cobination dose-response matrix. It's column name
and row name are representing the concerntrations of drug added to column 
and row, respectively. The values in matrix indicate the inhibition rate to
cell growth.}

\item{quiet}{A logical value. If it is \code{TRUE} then the warning message
will not show during calculation.}

\item{drug.col.model}{(optional) a character. It indicates the model used for
fitting dose-response curve for drug added to columns.}

\item{drug.row.model}{(optional) a character. It indicates the model type
used for fitting dose-response curve for drug added to rows.}
}
\value{
A matrix for Synergy score calculated via reference model introduced
  by Loewe, S.
}
\description{
\code{Loewe} calculates the synergy score matrix from a dose-response
matrix by using a druginteraction reference model introduced by Loewe in 1953.
}
\details{
Loewe model is a reference model for evaluating the combination 
effect of two drugs. The basic assumption of this model is "The referece 
effect of drug combination is the expected effect of a drug combined with 
itself". \cr
\cr
The optional arguments \code{drug.col.model}, \code{drug.row.model} are 
designed for reuse the single drug dose response model fitting results, 
if it has been down before. Functions \code{\link{FitDoseResponse}} and 
\code{\link{ExtractSingleDrug}} could be used to calculate these arguments.
}
\examples{
# No single drug fitted modle before
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
response.mat <- data$dose.response.mats[[1]]
Loewe.score <- Loewe(response.mat)

# Single drug dose response models have been fitted before.
drug.row.model <- FitDoseResponse(ExtractSingleDrug(response.mat, dim="row"))
drug.col.model <- FitDoseResponse(ExtractSingleDrug(response.mat, dim="col"))
Loewe.score <- Loewe(response.mat, drug.col.model=drug.col.model,
                     drug.row.model=drug.row.model)
}
\references{
\itemize{
   \item{Yadav B, Wennerberg K, Aittokallio T, Tang J.(2015).
   \href{https://doi.org/10.1016/j.csbj.2015.09.001}{Searching for Drug
   Synergy in Complex Dose-Response Landscape Using an Interaction Potency
   Model.}Comput Struct Biotechnol J, 13:504– 513.}
   \item{[Loewe, 1953] Loewe, S. (1953).
   \href{https://www.ncbi.nlm.nih.gov/pubmed/13081480}{The problem of
   synergism and antagonism of combined drugs.} Arzneimittelforschung,
   3(6):285–290.
   }
}
}
\author{
\itemize{
   \item{Liye He \email{liye.he@helsinki.fi}}
   \item{Jing Tang \email{jing.tang@helsinki.fi}}
   \item{Shuyu Zheng \email{shuyu.zheng@helsinki.fi}}
}
}
