% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_dose_response.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the drug combination dose-response data}
\usage{
PlotDoseResponse(data, save.file = FALSE)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}}

\item{save.file}{a parameter to specify if the visualization results are saved as pdf files in current working directory or not. If it is FALSE, the results are
returned as a list of the plots. It is FALSE by default.}
}
\value{
if save.file parameter is TRUE, pdf files are returned. Otherwise, the plots are only displayed.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponse(data)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

