% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_dose_response.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the drug combination dose-response data}
\usage{
PlotDoseResponse(data, adjusted = TRUE, save.file = FALSE,
  pair.index = NULL, ...)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{adjusted}{a logical value. If it is \code{FALSE}, original response 
matrix will be plotted. If it is \code{TRUE}, adjusted response matrix will
be plotted.}

\item{save.file}{a parameter to specify if the visualization results are 
saved as pdf files in current working directory or not. If it is FALSE, the 
results are returned as a list of the plots. It is FALSE by default.}

\item{pair.index}{a parameter to specify which drug combination if there are
many drug combinations in the data. By default, it is NULL so that the
visualization of all the drug combinations in the data is returned.}

\item{...}{further graphical parameters from \code{\link{plot}} for plotting 
the single drug dose-response curve. Use e.g., cex.lab to change the axis 
label size and cex.axis to change the tick size of axises.}
}
\value{
if save.file parameter is TRUE, pdf files are returned. Otherwise, 
  the plots are only displayed.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponse(data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi},
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  }
}
