% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_dose_response.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the drug combination dose-response data}
\usage{
PlotDoseResponse(
  data,
  adjusted = TRUE,
  pair.index = NULL,
  color.low.response = "green",
  color.point = "red",
  color.high.response = "red",
  color.conc = "red",
  save.file = FALSE,
  file.type = "pdf",
  file.name = NULL,
  ...
)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{adjusted}{a logical value. If it is \code{FALSE}, original response 
matrix will be plotted. If it is \code{TRUE}, adjusted response matrix will
be plotted.}

\item{pair.index}{a parameter to specify which drug combination if there are
many drug combinations in the data. By default, it is NULL so that the
visualization of all the drug combinations in the data is returned.}

\item{color.low.response}{a charactor in R color format. It indicates the 
color for the response lower than 0. Default setting is "green".}

\item{color.point}{a charactor in R color format. It indicates the 
color for the data point in the plot. Default setting is "red".}

\item{color.high.response}{a charactor in R color format. It indicates the 
color for the response higher than 0. Default setting is "red".}

\item{color.conc}{a charactor in R color format. It indicates the 
color for the concentrations label for axises. Default setting is "red".}

\item{save.file}{a parameter to specify if the visualization results are 
saved as pdf files in current working directory or not. If it is FALSE, 
the results are returned as a list of the plots. It is FALSE by default.}

\item{file.type}{a character. It indicates the format of files you want to 
save as. Default is "pdf". Available values are "jpeg", "bmp", "png", 
"tiff", "pdf", "svg".}

\item{file.name}{a character vector. It indicates the file names, if 
user chose to save the plot to local directory.If it is not defined by
user, a default name will be assigned.}

\item{...}{further graphical parameters from \code{plot} for
plotting the single drug dose-response curve. Use e.g., cex.lab to change 
the axis label size and cex.axis to change the tick size of axises.}
}
\value{
NULL. The plot will be saved into a local file if 
  \code{save.file = TRUE}. If \code{save.file = FALSE}, the plot will be 
  printed in default graphic device.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponse(data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi},
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  }
}
