% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_multiple_plots.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the Drug Combination Dose-response Data}
\usage{
PlotDoseResponse(
  data,
  block_ids = NULL,
  drugs = c(1, 2),
  adjusted = TRUE,
  statistic = NULL,
  summary_statistic = "mean",
  high_value_color = "#A90217",
  low_value_color = "#2166AC",
  point_color = "#C24B40",
  curve_color = "black",
  Emin = NA,
  Emax = NA,
  save_file = FALSE,
  file_type = "pdf",
  file_name = NULL,
  width = 12,
  height = 6
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{block_ids}{A vector of characters/integers or NULL. It contains the
block IDs for the blocks to visualize. By default, it is NULL so that the
visualization of all the drug combinations in input \code{data}.}

\item{drugs}{A vector of characters or integers with length of 2. It contains
the index for two drugs to plot. For example, \code{c(1, 2)} indicates to
plot "drug1" and "drug2" in the input \code{data}.}

\item{adjusted}{A logical value. If it is \code{FALSE}, original response
matrix will be plotted. If it is \code{TRUE}, adjusted response matrix will
be plotted.}

\item{statistic}{A character or NULL. It indicates the statistics printed
in the plot while there are replicates in input data. Available values are:
\itemize{
  \item \strong{sem} Standard error of mean;
  \item \strong{ci} 95% confidence interval.
}
If it is \code{NULL}, no statistics will be printed.}

\item{summary_statistic}{A vector of characters or NULL. It indicates the
summary statistics printed in heatmap for all the \code{plot_value} in
whole combination matrix. Available values are:
\itemize{
  \item \strong{mean} Median value for all the responses or synergy
  scores in the matrix;
  \item \strong{median} Median value for all the responses or synergy
  scores in the matrix;
  \item \strong{quantile_90} 90% quantile. User could change the number to
  print different sample quantile. For example quantile_50 equal to median. 
}
If it is \code{NULL}, no statistics will be printed.}

\item{high_value_color}{An R color value. It indicates the color for the
high values.}

\item{low_value_color}{An R color value. It indicates the color for low
values.}

\item{point_color}{An R color value. It indicates the color for points in
dose response curve plots.}

\item{curve_color}{An R color value. It indicates the color for curves in 
dose response curve plots.}

\item{Emin}{A numeric or \code{NA}. the minimal effect of the drug used in
the 4-parameter log-logistic function to fit the dose-response curve. If
it is not NA, it is fixed the value assigned by the user. Default setting
is \code{NA}.}

\item{Emax}{A numeric or \code{NA}. the maximal effect of the drug used in
the 4-parameter log-logistic function to fit the dose-response curve. If
it is not NA, it is fixed the value assigned by the user. Default setting
is \code{NA}.}

\item{save_file}{A parameter to specify if the visualization results are
saved as pdf files in current working directory or not. If it is FALSE,
the results are returned as a list of the plots. It is FALSE by default.}

\item{file_type}{A character. It indicates the format of files you want to
save as. Default is "pdf". Available values are "jpeg", "bmp", "png",
"tiff", "pdf", "svg".}

\item{file_name}{A character vector. It indicates the file names, if
user chose to save the plot to local directory.If it is not defined by
user, a default name will be assigned.}

\item{width}{a numeric value. It indicates the width of saved file.}

\item{height}{a numeric value. It indicates the height of saved file.}
}
\value{
A list of plot objects recorded by \link[grDevices]{recordPlot}. The 
  plot will be saved into a local file if \code{save.file = TRUE}. If 
  \code{save.file = FALSE}, the plot will be printed in default graphic
  device.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
\dontrun{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
plots <- PlotDoseResponse(data)
}
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi},
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  }
}
