% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dose_response.R
\name{PlotDoseResponse}
\alias{PlotDoseResponse}
\title{Visualize the drug combination dose-response data}
\usage{
PlotDoseResponse(
  data,
  adjusted = TRUE,
  block_ids = NULL,
  drugs = c(1, 2),
  statistic = NULL,
  title_color = "#635F5D",
  subtitle_color = "#635F5D",
  axis_text_color = "#635F5D",
  axis_title_color = "#635F5D",
  high_response_color = "#B2182B",
  low_response_color = "#2166AC",
  point_color = "#C20F00",
  curve_color = "black",
  save_file = FALSE,
  file_type = "pdf",
  file_name = NULL,
  width = 12,
  height = 6,
  ...
)
}
\arguments{
\item{data}{a list object generated by function \code{\link{ReshapeData}}.}

\item{adjusted}{a logical value. If it is \code{FALSE}, original response
matrix will be plotted. If it is \code{TRUE}, adjusted response matrix will
be plotted.}

\item{block_ids}{A vector characters or NULL. It contains the block IDs for
the blocks user want to visualize. By default, it is NULL so that the
visualization of all the drug combinations in the data is returned.}

\item{save_file}{a parameter to specify if the visualization results are
saved as pdf files in current working directory or not. If it is FALSE,
the results are returned as a list of the plots. It is FALSE by default.}

\item{file_type}{a character. It indicates the format of files you want to
save as. Default is "pdf". Available values are "jpeg", "bmp", "png",
"tiff", "pdf", "svg".}

\item{file_name}{a character vector. It indicates the file names, if
user chose to save the plot to local directory.If it is not defined by
user, a default name will be assigned.}

\item{width}{a numeric value. It indicates the width of saved file.}

\item{height}{a numeric value. It indicates the height of saved file.}

\item{...}{further graphical parameters from \code{plot} for
plotting the single drug dose-response curve. Use e.g., cex.lab to change
the axis label size and cex.axis to change the tick size of axises.}

\item{color.low.response}{a charactor in R color format. It indicates the
color for the response lower than 0. Default setting is "green".}

\item{color.high.response}{a charactor in R color format. It indicates the
color for the response higher than 0. Default setting is "red".}

\item{color.conc}{a charactor in R color format. It indicates the
color for the concentrations label for axises. Default setting is "red".}

\item{Point_color}{a charactor in R color format. It indicates the
color for the data point in the plot. Default setting is "red".}
}
\value{
NULL. The plot will be saved into a local file if
  \code{save.file = TRUE}. If \code{save.file = FALSE}, the plot will be
  printed in default graphic device.
}
\description{
A function to visualize the drug combination dose-response data
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
PlotDoseResponse(data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi},
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  }
}
