% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_drugs.R
\name{PlotMultiDrugSurface}
\alias{PlotMultiDrugSurface}
\title{3D Plot for Multi-drug Combination Dose-Response/Synergy Scores}
\usage{
PlotMultiDrugSurface(
  data,
  plot_block = 1,
  plot_value = "response",
  statistic = NULL,
  distance_method = "euclidean",
  high_value_color = "#A90217",
  low_value_color = "#2166AC",
  point_color = "#DDA137"
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{plot_block}{A character/integer. It indicates the block ID for the
block to visualize.}

\item{plot_value}{A vector of characters. It contains the name of one or more
metrics to be visualized. If the \code{data} is the direct output from 
\link{ReshapeData}, the values for this parameter are:
\itemze{
  \item \strong{response_origin} The original response value in input data.
  It might be % inhibition or % viability.
  \item \strong{response} The % inhibition after preprocess by function 
  \link{ReshapeData}
}
If the \code{data} is the output from \link{CalculateSynergy}, following
values are also available:
\itemze{
  \item \strong{ZIP_ref, Bliss_ref, HSA_ref, Loewe_ref} The reference
  additive effects predicted by ZIP, Bliss, HSA or Loewe model,
  respectively.
  \item \strong{ZIP_synergy, Bliss_synergy, HSA_synergy, Loewe_synergy}
  The synergy score calculated by ZIP, Bliss, HSA or Loewe model,
  respectively.
  \item \strong{ZIP_fit} The response fitted by ZIP model.
}}

\item{statistic}{A character or NULL. It indicates the statistics printed
in the plot while there are replicates in input data. Available values are:
\itemize{
  \item \strong{sd} Standard deviation;
  \item \strong{sem} Standard error of mean;
  \item \strong{CI95} 95% confidence interval.
}
If it is \code{NULL}, no statistics will be printed.}

\item{distance_method}{The methods to calculate the distance between
different data points from the concentration of drugs. The distance matrix
is used for dimension reduction. This parameter is used to set the 
parameter \code{method} for \link[vegdist]{vegan}. The default values is
"euclidean".}

\item{high_value_color}{An R color value. It indicates the color for the
high values.}

\item{low_value_color}{An R color value. It indicates the color for low
values.}

\item{point_color}{An R color value. It indicates the color for data points.}
}
\value{
A plotly plot object.
}
\description{
This function will generate a dynamic 3D plot response values or synergy
scores for all the observed data points in a multi-drug combination block.
The concentrations of drugs will be projected to 2 dimensions and plot along 
x an y axis. A surface for the selected \code{plot_value} and points for
all the concentration combinations will be plotted.
}
\examples{
data <- ReshapeData(NCAST_10023)
PlotMultiDrugSurface(data)
}
