% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitive_synergy.R
\name{PlotSensitiveSynergy}
\alias{PlotSensitiveSynergy}
\title{Plot Sensitive-Synergy Plot for All the Combinations in the Input Data}
\usage{
PlotSensitiveSynergy(
  data,
  plot_synergy,
  point_color = "#2D72AD",
  point_size = 1,
  point_label_color = "#2D72AD",
  show_labels = FALSE,
  plot_title = NULL,
  dynamic = FALSE,
  label_size = 10
)
}
\arguments{
\item{data}{A list object generated by function \code{\link{ReshapeData}}.}

\item{plot_synergy}{A character value. It indicates the synergy score for
visualization. The available values are: "ZIP", "HSA", Bliss", "Leowe".}

\item{point_color}{An R color value. It indicates the color for the points.}

\item{point_size}{A numeric value. It indicates the size of points. The unit
is "mm"}

\item{point_label_color}{An R color value. It indicates the color for the
label of data points.}

\item{show_labels}{A logic value. It indicates whether to show the labels
along with points or not.}

\item{dynamic}{A logical value. If it is \code{TRUE}, this function will
use \link[plotly]{plot_ly} to generate an interactive plot. If it is
\code{FALSE}, this function will use \link[lattice]{wireframe} to generate
a static plot.}

\item{label_size}{A numeric value. It controls the size of the labels in "pt"}
}
\value{
A ggplot object.
}
\description{
This function will generate a scatter plot for all the combinations in the
input data. The x-axis is the Combination Sensitive score (CSS).
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
data <- CalculateSynergy(data, method = c("ZIP"))
PlotSensitiveSynergy(data, plot_synergy = "ZIP")
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
