% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_synergy.R
\name{PlotSynergy}
\alias{PlotSynergy}
\title{Drug interaction landscape}
\usage{
PlotSynergy(
  data,
  type = "2D",
  save.file = FALSE,
  pair.index = NULL,
  len = 3,
  legend.start = NULL,
  legend.end = NULL,
  legend.lab.cex = 1,
  row.range = NULL,
  col.range = NULL,
  color.low.value = "green",
  main.title.cex = 1,
  axis.lab.cex = 0.8,
  axis.title.cex = 1,
  color.high.value = "red",
  file.name = NULL,
  file.type = "pdf",
  height = NULL,
  width = 12
)
}
\arguments{
\item{data}{a list object generated by function
\code{\link{CalculateSynergy}}.}

\item{type}{a parameter to specify the type of the interaction landscape, 2D,
3D or both. By default, 2D interaction landscape is returned.}

\item{save.file}{a logical parameter to specify if the interaction landscape
is saved as a pdf file in the current working directory or returned as an
R object. By default, it is FALSE.}

\item{pair.index}{a parameter to specify which drug combination if there are
many drug combinations in the data. By default, it is NULL so that the
synergy score visualization of all the drug combinations in the data is
returned.}

\item{len}{a parameter to specify how many values need to be predicted
between two concentrations. It is used to control the smoothness of the
synergy surface in the plot.}

\item{legend.start}{a parameter to specify the starting point of the legend.
By defualt, it is NULL so the legend starting point is fixed by the data
automatically.}

\item{legend.end}{a parameter to specify the ending point of the legend. By
defualt, it is NULL so the legend ending point is fixed by the data
automatically.}

\item{legend.lab.cex}{a numeric value. The magnification to be used for legend
labels.}

\item{row.range}{a parameter to specify the starting and ending concentration
of the drug on y-axis. Use e.g., c(1, 3) to specify that only from 1st to
3rd concentrations of the drug on y-axis are used. By default, it is NULl
so all the concentrations are used.}

\item{col.range}{a parameter to specify the starting and ending concentration
of the drug on x-axis. Use e.g., c(1, 3) to specify that only from 1st to
3rd concentrations of the drug on x-axis are used. By default, it is NULl
so all the concentrations are used.}

\item{color.low.value}{a charactor in R color format. It indicates the
color for the synergy score lower than 0. Default setting is "green".}

\item{main.title.cex}{a numeric value. The magnification to be used for axis
annotation.}

\item{axis.lab.cex}{a numeric value. The magnification to be used for x, y and z
axis labels.}

\item{axis.title.cex}{a numeric value. The magnification to be used for main titles.}

\item{color.high.value}{a charactor in R color format. It indicates the
color for the synergy score higher than 0. Default setting is "red".}

\item{file.name}{a character vector. It indicates the file names, if
user chose to save the plot to local directory.If it is not defined by
user, a default name will be assigned.}

\item{file.type}{a character. It indicates the format of files you want to
save as. Default is "pdf". Available values are "jpeg", "bmp", "png",
"tiff", "pdf", "svg".}

\item{height}{a numeric value. It indicates the height of the output plot.
If it is \code{NULL} (default setting), it will be automatically set according
to the \code{width} and \code{type} parameters. If the \code{type} is "all",
height = width, otherwise height = width/2.}

\item{width}{a numeric value. It indicates the width of the output plot.}
}
\value{
a numeric value. It is the summarized synergy score.
}
\description{
A function to visualize the synergy scores for drug combinations as 2D or 3D
interaction landscape over the dose-response matrix.
}
\examples{
data("mathews_screening_data")
data <- ReshapeData(mathews_screening_data)
scores <- CalculateSynergy(data)
PlotSynergy(scores, "2D")
PlotSynergy(scores, "3D", save.file = TRUE, file.name = c("plot1", "plot2"))
}
\author{
\itemize{
   \item{Liye He \email{liye.he@helsinki.fi}}
   \item{Shuyu Zheng \email{shuyu.zheng@helsinki.fi}}
}
}
