% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dose_response_model.R
\name{PredictModelSpecify}
\alias{PredictModelSpecify}
\title{Predict Response Value at Certain Drug Dose}
\usage{
PredictModelSpecify(model, dose)
}
\arguments{
\item{model}{An object for fitted model from drm::drc function L.4 or LL.4
model.}

\item{dose}{A numeric value. It indicates the concentration of drug at which
the response will be predicted.}
}
\value{
A numeric value. It is the response value of cell line to the drug at
inputted dose.
}
\description{
\code{PredictModelSpecify} uses \code{\link[drc]{drm}} function to fit the
dose response model and generate the predict response value at the given dose.
}
\details{
\strong{Note}: Random number generator used in \code{AddNoise} with
\code{method = "random"}. If the analysis requires reproducibility, please
set the random seed before calling this function.
}
\author{
\itemize{
  \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
  \item Jing Tang \email{jing.tang@helsinki.fi}
}
}
