% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reshape_data.R
\name{ReshapeData}
\alias{ReshapeData}
\title{Pre-processe the response data for furhter calculation and plot}
\usage{
ReshapeData(
  data,
  impute = TRUE,
  noise = TRUE,
  correction = "non",
  data.type = "viability"
)
}
\arguments{
\item{data}{drug combination response data in a data frame format}

\item{impute}{a logical value. If it is \code{TRUE}, the \code{NA} values
will be imputed by \code{\link{ImputeNA}}. Default is \code{TRUE}.}

\item{noise}{a logical value. It indicates whether or not adding noise to
to the "response" values in the matrix. Default is \code{TRUE}.}

\item{correction}{a character. This argument is extended from the argument 
\code{method} of \code{\link{CorrectBaseLine}} function. There are three 
available valuse: \code{non}, \code{part}, \code{all}. 
The default setting is \code{non}.}

\item{data.type}{a parameter to specify the response data type which can be 
either "viability" or "inhibition".}
}
\value{
a list of the following components:
  \itemize{
    \item \strong{dose.response.mats} a list of the dose-response matrices 
      with \%inhibition as the response data. Row names and column names are 
      drug concentrations.
  \item \strong{adjusted.response.mats} The dose response matrix adjusted.
    The processes are chosen by arguments \code{impute}, \code{noise}, and
    \code{correction}. If no process was chosen, the final result will not
    contain this result.
  \item \strong{drug.pairs} a data frame contains the name of the row drug,
    the name of the column drug, concentration unit and block IDs.
  }
}
\description{
A function to transform the response data from data frame format to 
dose-response matrices. Several processes could be chose to add noise, impute
missing values or correct base line to the dose-response matrix.
}
\details{
The input data must contain the following columns: block_id, 
  drug_row, drug_col, response, conc_r, conc_c, conc_r_unit, conc_c_unit.
}
\examples{
data("mathews_screening_data")
# set a random number seed for generating the noises
set.seed(1) 
data <- ReshapeData(mathews_screening_data)
}
\author{
\itemize{
    \item Liye He \email{liye.he@helsinki.fi}
    \item Shuyu Zheng \email{shuyu.zheng@helsinki.fi}
 }
}
